Fix byte conversion in HMAC output encoding

+++ core/service/ktorExt/src/iosMain/kotlin/com/maxrave/ktorext/crypto/Hmac.ios.kt
@@ -54,10 +54,10 @@ actual class Hmac actual constructor(algorithm: String, private val secretKey: S
                     )
                 }
             }
-        val bytes = ByteArray(digestLength) { i ->
-            output[i].toInt().toByte()   
-        }
-            Base64.encode(bytes)
+            val bytes = ByteArray(digestLength) { i ->
+                output[i].toByte()
+            }
+        Base64.encode(bytes)
         }
     }
